Move and test merge_file_names
