+++ nuc_processing/NucProcess.py
@@ -45,7 +45,7 @@ from subprocess import Popen, PIPE, call
 from functools import partial
 from .NucSvg import SvgDocument
 from .NucContactMap import nuc_contact_map
-from .common import open_file_r
+from .common import open_file_r, strip_ext, merge_file_names
 
 PROG_NAME = 'nuc_process'
 VERSION = '1.1.1'
@@ -66,7 +66,6 @@ MIN_READ_LEN = 20
 NUM_MAP_FASTAS = 10
 SCORE_TAG = re.compile(r'\sAS:i:(\S+)')
 SCORE_TAG_SEARCH = SCORE_TAG.search
-FILENAME_SPLIT_PATT = re.compile('[_\.]')
 NCC_FORMAT = '%s %d %d %d %d %s %s %d %d %d %d %s %d %d %d\n'
 LOG_FILE_PATH = None
 STAT_FILE_PATH = None
@@ -202,54 +201,6 @@ def write_sam_file(ncc_file_path, ref_sam_file_1, ref_sam_file_2):
        write('%s\n%s\n' % pair_sam_lines(line1[10:], line2[10:]))
 
    return sam_file_path
-
-
-def merge_file_names(file_path1, file_path2, sep='_'):
-
-  # same dir, need non truncated name
-
-  dir_name1, file_name1 = os.path.split(file_path1)
-  dir_name2, file_name2 = os.path.split(file_path2)
-
-  if dir_name1 != dir_name2:
-    msg = 'Attempt to merge file names for file from different directories'
-    raise Exception(msg)
-
-  file_root1, file_ext1 = os.path.splitext(file_name1)
-  file_root2, file_ext2 = os.path.splitext(file_name2)
-
-  if file_ext1 != file_ext2:
-    msg = 'Attempt to merge file names with different file extensions'
-    raise Exception(msg)
-
-  parts1 = FILENAME_SPLIT_PATT.split(file_root1)
-  parts2 = FILENAME_SPLIT_PATT.split(file_root2)
-  parts3 = []
-
-  n1 = len(parts1)
-  n2 = len(parts2)
-  n = max(n1, n2)
-
-  for i in range(n):
-
-    if (i < n1) and (i < n2):
-      a = parts1[i]
-      b = parts2[i]
-
-      parts3.append(a)
-      if a != b:
-        parts3.append(b)
-
-    elif i < n1:
-      parts3.append(parts1[i])
-    else:
-      parts3.append(parts2[i])
-
-  file_root3 = sep.join(parts3)
-
-  file_path3 = os.path.join(dir_name1, file_root3 + file_ext1)
-
-  return file_path3
 
 
 def remove_promiscuous(ncc_file, num_copies=1, keep_files=True, zip_files=False,
